breaking: rename core pkg to utilities
