+++ packages/overlays/package.json
@@ -41,7 +41,7 @@
   },
   "dependencies": {
     "@embroider/addon-shim": "^1.8.7",
-    "@frontile/core": "workspace:0.17.0-alpha.3",
+    "@frontile/utilities": "workspace:0.17.0-alpha.3",
     "@frontile/theme": "workspace:0.17.0-alpha.3",
     "@glint/template": "1.3.0",
     "ember-css-transitions": "^4.4.0",
