+++ packages/utilities/rollup.config.mjs
@@ -17,16 +17,12 @@ export default {
   plugins: [
     // These are the modules that users should be able to import from your
     // addon. Anything not listed here may get optimized away.
-    addon.publicEntrypoints([
-      'components/**/*.js',
-      'utils/**/*.js',
-      'template-registry.js'
-    ]),
+    addon.publicEntrypoints(['components/**/*.js', 'template-registry.js']),
 
     // These are the modules that should get reexported into the traditional
     // "app" tree. Things in here should also be in publicEntrypoints above, but
     // not everything in publicEntrypoints necessarily needs to go here.
-    addon.appReexports(['components/**/*.js', 'utils/**/*.js']),
+    addon.appReexports(['components/**/*.js']),
 
     // Follow the V2 Addon rules about dependencies. Your code can import from
     // `dependencies` and `peerDependencies` as well as standard Ember-provided
