+++ packages/utilities/src/components/collapsible.md
