+++ packages/utilities/src/utils/safe-styles.ts
