+++ test-app/tsconfig.json
@@ -5,8 +5,8 @@
     "paths": {
       "@frontile/notifications": ["../packages/notifications/addon"],
       "@frontile/notifications/*": ["../packages/notifications/addon/*"],
-      "@frontile/core": ["../packages/core/addon"],
-      "@frontile/core/*": ["../packages/core/addon/*"],
+      "@frontile/utilities": ["../packages/core/addon"],
+      "@frontile/utilities/*": ["../packages/core/addon/*"],
       "*": ["types/*"]
     }
   },
